change user id type

+++ teq-log/teq-log.go
@@ -2,6 +2,6 @@ package teq_log
 
 import "log"
 
-func PrintLog(userID int64, userName string, err string) {
+func PrintLog(userID int, userName string, err string) {
 	log.Println("[", userID, ":", userName, "] ", err)
 }
